% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloXML.R
\name{add_annot}
\alias{add_annot}
\title{Add annotation to phyloXML objects}
\usage{
add_annot(x, df)
}
\arguments{
\item{x}{A \code{phyloXML} or \code{phyloXML_phylogeny} object.}

\item{df}{A data.frame with a \code{name} column, to be matched with the nodes in the phylogeny.}
}
\value{
The modified \code{phyloXML} or \code{phyloXML_phylogeny} object.
}
\description{
Provide a data.frame with a \code{name} column matching the names of the clades in the phylogeny, and the new columns will be available in the layout for plotting.
}
\details{
If x is a \code{phyloXML} object, the annotation is added to all of its phylogenies.
}
\note{
If some annotation was already present, the new data.frame is
merged with the old one by the \code{name} column, keeping the existing
annotation. Use clear_annot() to remove all annotations.
}
