% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloXML.R
\name{add_sp_annot}
\alias{add_sp_annot}
\alias{add_recGene_annot}
\title{Add annotation to recPhyloXML objects}
\usage{
add_sp_annot(x, df)

add_recGene_annot(x, df)
}
\arguments{
\item{x}{An object of \code{recPhyloXML} class.}

\item{df}{A data.frame with a \code{name} column, to be matched with the nodes in the phylogeny.}
}
\description{
Provide a data.frame with a \code{name} column matching the names of the clades in the phylogeny, and the new columns will be available in the layout for plotting.
}
\details{
add_sp_annot() adds the annotation to the spTree while add_recGene_annot() to the recGeneTrees.
}
\note{
If some annotation was already present, the new data.frame is
merged with the old one by the \code{name} column, keeping the existing
annotation. Use clear_sp_annot() or clear_recGene_annot() to remove
all annotations.
}
