% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.data.frame.phyloXML_layout}
\alias{as.data.frame.phyloXML_layout}
\title{Convert phyloXML layout to data frame}
\usage{
\method{as.data.frame}{phyloXML_layout}(cl)
}
\arguments{
\item{cl}{A list representing the phyloXML layout. The list sholud have multiple nested lists under the element named 'children'.}
}
\value{
A data frame containing the concatenated elements of the phyloXML layout.
}
\description{
This function converts a phyloXML layout (a list structure) into a data frame.
}
\examples{
cl <- list(name = "root", value = 1, children = list(
              list(name = "child1", value = 2, children = list()),
              list(name = "child2", value = 3, children = list())))
as.data.frame.phyloXML_layout(cl)

}
