% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloXML.R
\name{decorate_clades}
\alias{decorate_clades}
\title{Decorate Clades with a Specific Attribute}
\usage{
decorate_clades(clade, attr, FUN, ...)
}
\arguments{
\item{clade}{A list representing a phylogenetic clade. Each node in the clade is expected to have a \code{clade} element containing its child nodes.}

\item{attr}{A character string representing the name of the attribute to be added to each node.}

\item{FUN}{A function that computes the value of the attribute for each node. The function should accept the clade as its first argument and any additional parameters through \code{...}.}

\item{...}{Additional parameters to be passed to \code{FUN}.}
}
\value{
The input \code{clade} list, with each node augmented by the computed attribute.
}
\description{
This function decorates each node in a clade with a new attribute computed by a user-defined function. It recursively applies the function to each child node in the clade.
}
\examples{
# Create a sample clade
clade <- example_phyloXML_clade()
# Define a function to compute an attribute
compute_attr <- function(clade, factor = 1) {
  length(clade$clade) * factor
}
# Decorate the clade with the new attribute
decorated_clade <- decorate_clades(clade, "node_count", compute_attr, factor = 2)

}
