% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloXML.R
\name{find_mrca}
\alias{find_mrca}
\title{Find the Most Recent Common Ancestor (MRCA)}
\usage{
find_mrca(clade, node_names)
}
\arguments{
\item{clade}{A list of class \code{phyloXML_clade} representing a
phylogenetic tree.}

\item{node_names}{A character vector containing the names of the
nodes for which the MRCA is to be found.}
}
\value{
The function returns the name of the MRCA of the specified
nodes, if found. If the MRCA is not found, the function returns \code{NA}
with an attribute "which_node_names" indicating the nodes that were
not found in the tree.
}
\description{
This function identifies the MRCA of a set of nodes. It traverses through the clades and returns the MRCA, if possible.
}
\examples{
clade <- example_phyloXML_clade()
find_mrca(clade, c("B. subtilis", "E. coli"))
find_mrca(clade, c("B. subtilis", "Bogus"))

}
