% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloXML.R
\name{flip_recGene_children}
\alias{flip_recGene_children}
\title{Flip the children of a clade in recGeneTrees}
\usage{
flip_recGene_children(x, name, perm = NULL)
}
\arguments{
\item{x}{A \code{recPhyloXML} object.}

\item{name}{A character string specifying the name of the clade whose children are to be flipped.}

\item{perm}{An optional numeric vector specifying the new permutation order of the children. If \code{NULL}, the next permutation in lexycographic order is used.}
}
\value{
The \code{recPhyloXML} object with the modified recGeneTrees.
}
\description{
This function permutes the order of child clades. Every time it is called on the same parent clade, the order of its children becomes the next permutation in lexycographic order, coming back to the initial values when all possibilities have been exhausted. For example, for a node with three children, the initial order will be c(1, 2, 3); calling flip_children() once will give c(1, 3, 2), calling it again c(2, 1, 3), then c(2, 3, 1), c(3, 1, 2), c(3, 2, 1), and finally it will come back to c(1, 2, 3). The optional argument \code{perm} allows to specify a target permutation directly.
}
\details{
All of the recGeneTrees phylogenies will be affected, if they have a clade of the specified name.
}
\note{
Flipping the order of children does NOT change the topology of the tree. It matters only for visualization purposes.
}
\examples{
phylogeny <- example_phyloXML_phylogeny()
flipped_phylogeny <- flip_children(phylogeny, "Bacteria")
}
