% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloXML.R
\name{import_branch_lengths}
\alias{import_branch_lengths}
\title{Import Branch Lengths from a \code{phylo} Object}
\usage{
import_branch_lengths(x, phylo)
}
\arguments{
\item{x}{Either a \code{phyloXML_phylogeny} or a \code{recPhyloXML} object to which the branch lengths are to be imported.}

\item{phylo}{A \code{phylo} object containing the branch lengths.}
}
\value{
The modified phylogeny with the branch lengths from the \code{phylo} object.
}
\description{
This function takes branch lengths from an \code{ape::phylo}
object and imports them into the corresponding nodes of a
\code{phyloXML_phylogeny} or \code{recPhyloXML} object. For a \code{recPhyloXML},
the branch lengths are imported only to the species tree.
}
\details{
The \code{edge.length} and \code{node.label} fields of the \code{phylo} object must exist.
}
\examples{
\dontrun{
phyloxml <- example_phyloXML()
ape_tree <- ape::read.tree("path/to/tree.nwk")
phyloxml[[1]] <- import_branch_lengths(phyloxml[[1]], ape_tree)
}

\dontrun{
recphyloxml <- example_recPhyloXML()
ape_tree <- ape::read.tree("path/to/tree.nwk")
recphyloxml <- import_branch_lengths(recphyloxml, ape_tree)
}

}
