% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_layout}
\alias{merge_layout}
\title{Merge Phylogeny Data with Layout Data}
\usage{
merge_layout(phylogeny, layout)
}
\arguments{
\item{phylogeny}{A list of class \verb{phyloXML_phylogeny, potentially annotated with an }annot` attribute.}

\item{layout}{A data frame representing the layout data.}
}
\value{
A data frame resulting from the merge of the phylogeny and layout data frames, with annotation data if available.
}
\description{
This function merges a \code{phyloXML_phylogeny} list with a layout data frame based on a common column 'name'.
Additionally, it merges annotation data if present in the phylogeny.
}
\examples{
phylogeny <- example_phyloXML_phylogeny()
layout <- data.frame(name = c("unnamed", "Octopus", "Bacteria", "E. coli", "B. subtilis"), position = c(100, 100, 200, 150, 250))
attr(phylogeny, "annot") <- data.frame(name = c("Octopus"), annotation = c("annot1"))
merge_layout(phylogeny, layout)
}
