% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{phyloXML_examples}
\alias{phyloXML_examples}
\alias{example_recPhyloXML_file}
\alias{example_phyloXML_file}
\alias{example_recPhyloXML}
\alias{example_phyloXML}
\alias{example_phyloXML_phylogeny}
\alias{example_phyloXML_clade}
\title{Get examples for recPhyloXML or phyloXML objects}
\usage{
example_recPhyloXML_file(name = NULL)

example_phyloXML_file(name = NULL)

example_recPhyloXML()

example_phyloXML()

example_phyloXML_phylogeny()

example_phyloXML_clade()
}
\value{
An example of the requested type.
}
\description{
Get examples for recPhyloXML or phyloXML objects
}
