% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloXML.R
\name{read_recPhyloXML}
\alias{read_recPhyloXML}
\title{Read a recPhyloXML file}
\usage{
read_recPhyloXML(x, ...)
}
\arguments{
\item{x}{A string, a connection, or a raw vector (see xml2::read_xml).}

\item{...}{Additional arguments passed on to xml2::read_xml.}
}
\value{
A list with two elements: spTree and recGeneTrees. spTree is a simple phylogeny, recGeneTrees is a list of phylogenies.
}
\description{
Read a reconciliated tree from a file in recPhyloXML format and parse it
into a list of spTree and recGeneTrees objects.
}
\note{
The \verb{<name>} of the nodes in the tree will be made unique, if necessary, by adding
suffixes of the form \verb{#1}, \verb{#2}, and so on. Unnamed clades will be given
"unnamed" as their name. This is because the functions in this package
heavily rely on the clades being uniquely named.
}
\section{Phylogeny objects}{

Each phylogeny is a list of class \code{phyloXML_phylogeny} with one field
for each attribute or child of the \verb{<phylogeny>} tag. The \code{clade} field
is a nested list structure of class \code{phyloXML_clade}, with one field
for each attribute of child of the \verb{<clade>} tag.
}

\examples{
path_to_xml <- example_phyloXML_file()
read_phyloXML(path_to_xml)

}
