% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recphylo_example}
\alias{recphylo_example}
\title{Get the path to an example RecPhyloXML}
\usage{
recphylo_example(name = NULL)
}
\arguments{
\item{name}{The name of the example.}
}
\value{
If name is \code{NULL}, a vector of possible examples. Otherwise, the
path to the requested example.
}
\description{
Get the path to an example RecPhyloXML
}
