% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_phyloXML}
\alias{write_phyloXML}
\title{Save a "phylo" object to a file in phyloXML format}
\usage{
write_phyloXML(tree, path)
}
\arguments{
\item{tree}{A phylogenetic tree (e.g. from ape::read.tree()).}

\item{path}{Path to the file where the tree will be written.}
}
\description{
Save a "phylo" object to a file in phyloXML format
}
