% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recPhyloParse.R
\name{RecPhylo}
\alias{RecPhylo}
\title{R6 Class RecPhylo}
\value{
A new RecPhylo object.
}
\description{
Class used to import and manipulate \href{https://en.wikipedia.org/wiki/Phylogenetic_reconciliation}{reconciliated phylogenetic trees}
from \href{http://phylariane.univ-lyon1.fr/recphyloxml/}{RecPhyloXML} files.
}
\examples{

## ------------------------------------------------
## Method `RecPhylo$new`
## ------------------------------------------------

xml_file <- system.file("extdata", "example_1.recphyloxml", package = "recPhyloParse")
rp <- RecPhylo$new(xml_file)

## ------------------------------------------------
## Method `RecPhylo$redraw`
## ------------------------------------------------

xml_file <- system.file("extdata", "example_1.recphyloxml", package = "recPhyloParse")
rp <- RecPhylo$new(xml_file)
rp$redraw(x_padding = 5)

## ------------------------------------------------
## Method `RecPhylo$import_branch_lengths`
## ------------------------------------------------

xml_file <- system.file("extdata", "example_1.recphyloxml", package = "recPhyloParse")
rp <- RecPhylo$new(xml_file)
rp$redraw(x_padding = 5)
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{spList}}{Species tree in list form.}

\item{\code{spNodes}}{data.frame with the nodes of the species tree.}

\item{\code{spEdges}}{data.frame with the edges of the species tree.}

\item{\code{recGeneList}}{Genes tree in list form.}

\item{\code{recGeneNodes}}{data.frame with the nodes of the genes tree.}

\item{\code{recGeneEdges}}{data.frame with the edges of the genes tree.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RecPhylo-new}{\code{RecPhylo$new()}}
\item \href{#method-RecPhylo-redraw}{\code{RecPhylo$redraw()}}
\item \href{#method-RecPhylo-import_branch_lengths}{\code{RecPhylo$import_branch_lengths()}}
\item \href{#method-RecPhylo-print}{\code{RecPhylo$print()}}
\item \href{#method-RecPhylo-summary}{\code{RecPhylo$summary()}}
\item \href{#method-RecPhylo-plot}{\code{RecPhylo$plot()}}
\item \href{#method-RecPhylo-clone}{\code{RecPhylo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecPhylo-new"></a>}}
\if{latex}{\out{\hypertarget{method-RecPhylo-new}{}}}
\subsection{Method \code{new()}}{
Create a new RecPhylo object from a RecPhyloXML file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecPhylo$new(
  xml_file,
  use_branch_length = "branch_length",
  x_padding = 1,
  branch_length_scale = 1,
  use_y_shift = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml_file}}{Path to the RecPhyloXML file.}

\item{\code{use_branch_length}}{What to use as branch length}

\item{\code{x_padding}}{Distance between species in the plot}

\item{\code{branch_length_scale}}{Multiplier for the branch length}

\item{\code{use_y_shift}}{Whether to apply the y_shift correction (can make the plot look prettier)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{xml_file <- system.file("extdata", "example_1.recphyloxml", package = "recPhyloParse")
rp <- RecPhylo$new(xml_file)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecPhylo-redraw"></a>}}
\if{latex}{\out{\hypertarget{method-RecPhylo-redraw}{}}}
\subsection{Method \code{redraw()}}{
Redraw the reconciliated tree, changing parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecPhylo$redraw(x_padding, use_branch_length, branch_length_scale, use_y_shift)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x_padding}}{Distance between species in the plot}

\item{\code{use_branch_length}}{What to use as branch length}

\item{\code{branch_length_scale}}{Multiplier for the branch length}

\item{\code{use_y_shift}}{Whether to apply the y_shift correction (can make the plot look prettier)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{xml_file <- system.file("extdata", "example_1.recphyloxml", package = "recPhyloParse")
rp <- RecPhylo$new(xml_file)
rp$redraw(x_padding = 5)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecPhylo-import_branch_lengths"></a>}}
\if{latex}{\out{\hypertarget{method-RecPhylo-import_branch_lengths}{}}}
\subsection{Method \code{import_branch_lengths()}}{
Import branch lengths from another species tree.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecPhylo$import_branch_lengths(phylo)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phylo}}{Object of class "phylo" (e.g. from ape's read.tree())}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{xml_file <- system.file("extdata", "example_1.recphyloxml", package = "recPhyloParse")
rp <- RecPhylo$new(xml_file)
rp$redraw(x_padding = 5)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecPhylo-print"></a>}}
\if{latex}{\out{\hypertarget{method-RecPhylo-print}{}}}
\subsection{Method \code{print()}}{
Print config, spNodes, and recGeneNodes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecPhylo$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecPhylo-summary"></a>}}
\if{latex}{\out{\hypertarget{method-RecPhylo-summary}{}}}
\subsection{Method \code{summary()}}{
Print summary information about the tree.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecPhylo$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecPhylo-plot"></a>}}
\if{latex}{\out{\hypertarget{method-RecPhylo-plot}{}}}
\subsection{Method \code{plot()}}{
Create a test plot of the tree.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecPhylo$plot()}\if{html}{\out{</div>}}
}

\subsection{Details}{
This method is used for testing purposes, just to check that all the
elements are OK. Feel free to copy the code and edit it to make your
own pretty plots.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecPhylo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RecPhylo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecPhylo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
